pyproject.toml
[project]
name = "thucchienai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.46.0",
    "openai>=2.6.0",
    "pillow>=12.0.0",
    "playwright>=1.55.0",
    "python-dotenv>=1.1.1",
]
